+++ UIFragmentsJorge/app/src/main/java/com/uc3m/foodstuff/api/ApiClient.kt
@@ -1,0 +1,36 @@
+package com.uc3m.foodstuff.api
+
+import com.google.gson.Gson
+import com.google.gson.GsonBuilder
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+
+class ApiClient {
+    private val BASE_URL: String = "https://reecipe-weekly-default-rtdb.europe-west1.firebasedatabase.app"
+
+    private val gson : Gson by lazy {
+        GsonBuilder().setLenient().create()
+    }
+
+    private val interceptor : HttpLoggingInterceptor by lazy {
+        HttpLoggingInterceptor().setLevel(HttpLoggingInterceptor.Level.BODY)
+    }
+
+    private val httpClient : OkHttpClient by lazy {
+        OkHttpClient.Builder().addInterceptor(interceptor).build()
+    }
+
+    private val retrofit : Retrofit by lazy {
+        Retrofit.Builder()
+                .baseUrl(BASE_URL)
+                .client(httpClient)
+                .addConverterFactory(GsonConverterFactory.create(gson))
+                .build()
+    }
+
+    val apiService: ApiService by lazy {
+        retrofit.create(ApiService::class.java)
+    }
+}
