+++ UIFragmentsJorge/app/src/main/java/com/uc3m/foodstuff/data/model/LoggedInUser.kt
@@ -1,0 +1,9 @@
+package com.uc3m.foodstuff.data.model
+
+/**
+ * Data class that captures user information for logged in users retrieved from LoginRepository
+ */
+data class LoggedInUser(
+        val userId: String,
+        val displayName: String
+)
