+++ UIFragmentsJorge/app/src/main/java/com/uc3m/foodstuff/ui/login/LoggedInUserView.kt
@@ -1,0 +1,9 @@
+package com.uc3m.foodstuff.ui.login
+
+/**
+ * User details post authentication that is exposed to the UI
+ */
+data class LoggedInUserView(
+        val displayName: String
+        //... other data fields that may be accessible to the UI
+)
