+++ UIFragmentsJorge/app/src/main/java/com/uc3m/foodstuff/ui/search/SearchFragment.kt
@@ -1,0 +1,31 @@
+package com.uc3m.foodstuff.ui.search
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.TextView
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.Observer
+import androidx.lifecycle.ViewModelProvider
+import com.uc3m.foodstuff.R
+
+class SearchFragment : Fragment() {
+
+    private lateinit var searchViewModel: SearchViewModel
+
+    override fun onCreateView(
+            inflater: LayoutInflater,
+            container: ViewGroup?,
+            savedInstanceState: Bundle?
+    ): View? {
+        searchViewModel =
+                ViewModelProvider(this).get(SearchViewModel::class.java)
+        val root = inflater.inflate(R.layout.fragment_search, container, false)
+        val textView: TextView = root.findViewById(R.id.text_notifications)
+        searchViewModel.text.observe(viewLifecycleOwner, Observer {
+            textView.text = it
+        })
+        return root
+    }
+}
