+++ UIFragmentsJorge/app/src/main/java/com/uc3m/foodstuff/ui/search/SearchViewModel.kt
@@ -1,0 +1,13 @@
+package com.uc3m.foodstuff.ui.search
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+
+class SearchViewModel : ViewModel() {
+
+    private val _text = MutableLiveData<String>().apply {
+        value = "This is search Fragment"
+    }
+    val text: LiveData<String> = _text
+}
