+++ UIFragmentsJorge/app/src/main/java/com/uc3m/foodstuff/ui/weekly/WeeklyViewModel.kt
@@ -1,0 +1,13 @@
+package com.uc3m.foodstuff.ui.weekly
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+
+class WeeklyViewModel : ViewModel() {
+
+    private val _text = MutableLiveData<String>().apply {
+        //value = "This is weekly Fragment"
+    }
+    val text: LiveData<String> = _text
+}
