--- UIFragmentsJorge/app/src/main/java/com/uc3m/uifragments/data/LoginDataSource.kt
@@ -1,24 +1,0 @@
-package com.uc3m.uifragments.data
-
-import com.uc3m.uifragments.data.model.LoggedInUser
-import java.io.IOException
-
-/**
- * Class that handles authentication w/ login credentials and retrieves user information.
- */
-class LoginDataSource {
-
-    fun login(username: String, password: String): Result<LoggedInUser> {
-        try {
-            // TODO: handle loggedInUser authentication
-            val fakeUser = LoggedInUser(java.util.UUID.randomUUID().toString(), "Jane Doe")
-            return Result.Success(fakeUser)
-        } catch (e: Throwable) {
-            return Result.Error(IOException("Error logging in", e))
-        }
-    }
-
-    fun logout() {
-        // TODO: revoke authentication
-    }
-}
