--- UIFragmentsJorge/app/src/main/java/com/uc3m/uifragments/db/AppDatabase.kt
@@ -1,27 +1,0 @@
-package com.uc3m.uifragments.db
-
-import android.content.Context
-import androidx.room.Database
-import androidx.room.Room
-import androidx.room.RoomDatabase
-
-@Database(entities = [Recipe::class], version = 1)
-abstract class AppDatabase : RoomDatabase() {
-    abstract fun recipeDao() : RecipeDao
-
-    companion object {
-        @Volatile
-        private var INSTANCE: AppDatabase? = null
-        fun getInstance(context: Context): AppDatabase {
-            return INSTANCE ?: synchronized(this) {
-                val instance = Room.databaseBuilder(
-                        context.applicationContext,
-                        AppDatabase::class.java,
-                        "food-stuff-db"
-                ).fallbackToDestructiveMigration().build()
-                INSTANCE = instance
-                instance
-            }
-        }
-    }
-}
