--- UIFragmentsJorge/app/src/main/java/com/uc3m/uifragments/db/RecipeDao.kt
@@ -1,22 +1,0 @@
-package com.uc3m.uifragments.db
-
-import androidx.room.Dao
-import androidx.room.Delete
-import androidx.room.Insert
-import androidx.room.Query
-
-@Dao
-interface RecipeDao {
-    @Query("SELECT * FROM recipe")
-    fun getAll(): List<Recipe>
-
-    @Query("SELECT * FROM recipe WHERE id LIKE :id")
-    fun getById(id: String): Recipe
-    // TODO do more queries to find other stuff
-
-    @Insert
-    fun insert(recipe: Recipe)
-
-    @Delete
-    fun delete(recipe: Recipe)
-}
