--- UIFragmentsJorge/app/src/main/java/com/uc3m/uifragments/ui/login/LoginFormState.kt
@@ -1,8 +1,0 @@
-package com.uc3m.uifragments.ui.login
-
-/**
- * Data validation state of the login form.
- */
-data class LoginFormState(val usernameError: Int? = null,
-                          val passwordError: Int? = null,
-                          val isDataValid: Boolean = false)
