--- UIFragmentsJorge/app/src/main/java/com/uc3m/uifragments/ui/profile/ProfileFragment.kt
@@ -1,34 +1,0 @@
-package com.uc3m.uifragments.ui.profile
-
-import android.content.Context
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.TextView
-import androidx.fragment.app.Fragment
-import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModelProvider
-import androidx.room.Room
-import com.uc3m.uifragments.R
-import com.uc3m.uifragments.db.AppDatabase
-
-class ProfileFragment : Fragment() {
-
-    private lateinit var profileViewModel: ProfileViewModel
-
-    override fun onCreateView(
-            inflater: LayoutInflater,
-            container: ViewGroup?,
-            savedInstanceState: Bundle?
-    ): View? {
-        profileViewModel =
-                ViewModelProvider(this).get(ProfileViewModel::class.java)
-        val root = inflater.inflate(R.layout.fragment_profile, container, false)
-        val textView: TextView = root.findViewById(R.id.text_username)
-        profileViewModel.text.observe(viewLifecycleOwner, Observer {
-            textView.text = it
-        })
-        return root
-    }
-}
