--- UIFragmentsJorge/app/src/main/java/com/uc3m/uifragments/ui/profile/ProfileViewModel.kt
@@ -1,13 +1,0 @@
-package com.uc3m.uifragments.ui.profile
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class ProfileViewModel : ViewModel() {
-
-    private val _text = MutableLiveData<String>().apply {
-        value = "Hello, Username."
-    }
-    val text: LiveData<String> = _text
-}
