+++ UIFragmentsJorge/app/src/main/res/navigation/mobile_navigation.xml
@@ -7,19 +7,19 @@
 
     <fragment
         android:id="@+id/navigation_profile"
-        android:name="com.uc3m.uifragments.ui.profile.ProfileFragment"
+        android:name="com.uc3m.foodstuff.ui.profile.ProfileFragment"
         android:label="@string/title_profile"
         tools:layout="@layout/fragment_profile" />
 
     <fragment
         android:id="@+id/navigation_weekly"
-        android:name="com.uc3m.uifragments.ui.weekly.WeeklyFragment"
+        android:name="com.uc3m.foodstuff.ui.weekly.WeeklyFragment"
         android:label="@string/title_weekly"
         tools:layout="@layout/fragment_weekly" />
 
     <fragment
         android:id="@+id/navigation_search"
-        android:name="com.uc3m.uifragments.ui.search.SearchFragment"
+        android:name="com.uc3m.foodstuff.ui.search.SearchFragment"
         android:label="@string/title_search"
         tools:layout="@layout/fragment_search" />
 </navigation>
